media: sunxi: sun6i-csi: Add support of MIPI CSI-2 for A83T
This patch add the support only for the Allwinner A83T MIPI CSI-2.
Currently, the driver does not support the  V3s MIPI CSI-2 controller.
On the A83T, the CSI controller is the same as the other V3s Soc, but
the MIPI CSI2 controller is not.

It was tested with the ov8865 image sensor.

Signed-off-by: Kévin L'hôpital <kevin.lhopital@bootlin.com>